// src/shared/api/generated/types/RecommendationConfig.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { FieldFilter } from './FieldFilter.ts'
import type { ScheduleDate } from './ScheduleDate.ts'

/**
 * @description Конфигурация рекомендаций
 * @example [object Object]
 */
export type RecommendationConfig = {
  /**
   * @description ID конфигурации
   */
  id?: string | null
  created_at?: string | null
  updated_at?: string | null
  /**
   * @description Название конфигурации
   * @type string
   */
  name: string
  /**
   * @description Описание конфигурации
   * @type string
   */
  description: string
  /**
   * @description Активна ли конфигурация
   * @default true
   * @type boolean | undefined
   */
  is_active?: boolean
  /**
   * @description Фильтры по полям тайтлов
   * @type array
   */
  title_field_filters: FieldFilter[]
  /**
   * @description Расписания запуска
   * @type array
   */
  schedules_dates: ScheduleDate[]
}